+++ src/fn/fn-ramp.js
@@ -60,6 +60,10 @@ var strategy = {
 
   exact: createSplitStrategy(function exactSelector (column, value) {
     return Number.isFinite(value) ? '[ ' + column + ' = ' + value + ' ]' : '[ ' + column + ' = "' + value + '" ]';
+  }),
+
+  '=': createSplitStrategy(function exactSelector (column, value) {
+    return Number.isFinite(value) ? '[ ' + column + ' = ' + value + ' ]' : '[ ' + column + ' = "' + value + '" ]';
   })
 };